rename styles.ts files
