+++ packages/client/ui-styles/src/index.ts
@@ -1,6 +1,6 @@
-export * from './common.styles';
+export * from './global-styles';
 export * from './tokens';
-export * from './text.styles';
+export * from './utility-styles';
 export * from './Text';
 export * from './SkeletonTxt';
 export * from './Header';
